requirements.txt
# 核心依赖
numpy>=1.21.0
matplotlib>=3.5.0
streamlit>=1.10.0
networkx>=2.6.0  # 图算法库，后续可能用于复杂的图分析

# 机器学习依赖 (Phase 2 预留)
torch>=1.12.0    # PyTorch
pandas>=1.3.0    # 数据处理
scikit-learn>=1.0 # 基础评估指标

pyyaml
